requirements.txt
Pillow==9.2.0
python-telegram-bot==20.0a4
